refactor(search): replace sidecar embedding with native ONNX
- Remove embedding_sidecar and platform-specific files
- Add native embedding implementation
- Add ONNX runtime integration
- Simplify embed sidecar runtime and main entry
- Remove main_nonwindows build constraint file
