--- internal/search/embedding_sidecar_nonwindows.go
@@ -1,7 +1,0 @@
-//go:build !windows
-
-package search
-
-import "os/exec"
-
-func configureSidecarCommand(cmd *exec.Cmd) {}
