--- internal/search/embedding_sidecar_windows.go
@@ -1,15 +1,0 @@
-//go:build windows
-
-package search
-
-import (
-	"os/exec"
-	"syscall"
-)
-
-func configureSidecarCommand(cmd *exec.Cmd) {
-	cmd.SysProcAttr = &syscall.SysProcAttr{
-		HideWindow:    true,
-		CreationFlags: 0x08000000,
-	}
-}
